#6034 – Combine two editors in ketcher-react package in order to make macro switcher enabled by default (#6265)
* #6034 – Combine two editors in ketcher-react package in order to make macro switcher enabled by default

* #6034 – Add loader when switching modes, change disable prop name, adjust readme files
