--- example-separate-editors/.eslintignore
@@ -1,4 +1,0 @@
-dist/
-public/
-build/
-config/
