--- example-separate-editors/.stylelintignore
@@ -1,2 +1,0 @@
-build/
-dist/
