--- example-separate-editors/README.md
@@ -1,6 +1,0 @@
-This example was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
-
-It is linked to the @ketcher/react package in the parent directory for development purposes.
-
-To run application for development purposes, please check instructions:
-[DEVNOTES.md](https://github.com/epam/ketcher/blob/master/DEVNOTES.md)
