--- example-separate-editors/package.json
@@ -1,94 +1,0 @@
-{
-  "name": "example-separate-editors",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "dev:standalone": "cross-env MODE=standalone vite",
-    "dev:standalone:macro": "cross-env ENABLE_POLYMER_EDITOR=true MODE=standalone vite",
-    "dev:remote": "cross-env MODE=remote vite",
-    "start:standalone": "cross-env MODE=standalone react-app-rewired start",
-    "start:remote": "cross-env MODE=remote react-app-rewired start",
-    "init:build": "shx mkdir -p build",
-    "init:dist": "cross-env-shell shx mkdir -p dist/$MODE",
-    "delete:dist": "cross-env-shell shx rm -rf dist/$MODE",
-    "copy:build": "cross-env-shell shx cp -r build/. dist/$MODE",
-    "prebuild": "cross-env-shell run-s init:build delete:dist init:dist",
-    "postbuild": "cross-env-shell npm run copy:build",
-    "build:remote": "cross-env-shell MODE=remote run-s prebuild build:react postbuild",
-    "build:standalone": "cross-env-shell MODE=standalone run-s prebuild build:react postbuild",
-    "build:react": "cross-env-shell MODE=$MODE react-app-rewired build && echo $MODE",
-    "build:react:analyze": "react-app-rewired build --analyze",
-    "build": "npm run build:standalone && npm run build:remote",
-    "build:standalone:analyze": "cross-env MODE=standalone run-s prebuild build:react:analyze postbuild",
-    "build:remote:analyze": "cross-env MODE=remote run-s prebuild build:react:analyze postbuild",
-    "test": "run-s test:prettier test:stylelint test:eslint:quiet test:types test:unit",
-    "test:eslint": "eslint . --ext .ts,.js,.jsx,.tsx",
-    "test:eslint:quiet": "npm run test:eslint -- --quiet",
-    "test:eslint:fix": "npm run test:eslint -- --fix",
-    "test:types": "tsc --noEmit",
-    "test:unit": "react-app-rewired test --passWithNoTests --watchAll=false",
-    "test:stylelint": "stylelint \"./**/*.{css,less}\" --formatter",
-    "test:prettier": "prettier --check \"./**/*.{js,ts,jsx,tsx,json}\"",
-    "prettier:write": "prettier --write \"./**/*.{js,jsx,json,ts,tsx}\"",
-    "stylelint:fix": "stylelint \"./**/*.{css,less}\" --formatter verbose --fix",
-    "serve:remote": "serve -s dist/remote -l 4001",
-    "serve:standalone": "serve -s dist/standalone -l 4002",
-    "eject": "react-scripts eject"
-  },
-  "dependencies": {
-    "ketcher-core": "*",
-    "ketcher-macromolecules": "*",
-    "ketcher-react": "*",
-    "ketcher-standalone": "*",
-    "normalize.css": "^8.0.1",
-    "react": "^18.2.0",
-    "react-app-polyfill": "^2.0.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "@rollup/plugin-replace": "^5.0.2",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^27.0.3",
-    "@types/node": "^16.11.12",
-    "@types/react": "^18.0.25",
-    "@types/react-dom": "^18.0.8",
-    "@vitejs/plugin-react": "^4.0.0",
-    "@welldone-software/why-did-you-render": "^4.3.1",
-    "copy-webpack-plugin": "^12.0.2",
-    "cross-env": "^7.0.3",
-    "customize-cra": "^1.0.0",
-    "eslint": "^8.44.0",
-    "git-revision-webpack-plugin": "^3.0.6",
-    "html-replace-webpack-plugin": "^2.5.6",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.5.1",
-    "react-app-rewired": "^2.2.1",
-    "react-scripts": "^5.0.1",
-    "serve": "^13.0.2",
-    "shx": "^0.3.3",
-    "source-map-loader": "^3.0.1",
-    "stylelint": "13.13.1",
-    "typescript": "^4.5.2",
-    "vite": "^4.5.5",
-    "vite-plugin-html": "^3.2.0",
-    "vite-plugin-raw": "^1.0.3",
-    "vite-plugin-svgr": "^2.4.0",
-    "webpack-bundle-analyzer": "^4.8.0"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "ie 11",
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "config-overrides-path": "config/webpack.config.js"
-}
