--- example-separate-editors/public/manifest.json
@@ -1,15 +1,0 @@
-{
-  "short_name": "Ketcher",
-  "name": "Ketcher",
-  "icons": [
-    {
-      "src": "favicon.ico",
-      "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/x-icon"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
