--- example-separate-editors/src/PolymerToggler/PolymerToggler.module.css
@@ -1,65 +1,0 @@
-/****************************************************************************
- * Copyright 2021 EPAM Systems
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- ***************************************************************************/
-
-.switch {
-  position: fixed;
-  display: block;
-  width: 24px;
-  height: 14px;
-  right: 50px;
-  bottom: 50px;
-}
-
-.switch input {
-  display: none;
-}
-
-.slider {
-  position: absolute;
-  cursor: pointer;
-  right: 0;
-  bottom: 0;
-  background-color: #c4c4c4;
-  -webkit-transition: 0.4s;
-  transition: background 0.4s;
-  border-radius: 7px;
-  width: 26px;
-  height: 14px;
-}
-
-.slider::before {
-  position: absolute;
-  content: '';
-  height: 10px;
-  width: 10px;
-  left: 2px;
-  top: 2px;
-  background-color: #343434;
-  -webkit-transition: 0.4s;
-  transition: transform 0.4s ease-in 0s;
-  border-radius: 50%;
-}
-
-.switch input:checked + .slider {
-  background-color: #343434;
-}
-
-.switch input:checked + .slider::before {
-  background-color: #fff;
-  -webkit-transform: translateX(12px);
-  -ms-transform: translateX(12px);
-  transform: translateX(12px);
-}
