--- example-separate-editors/src/PolymerToggler/PolymerToggler.tsx
@@ -1,34 +1,0 @@
-/****************************************************************************
- * Copyright 2021 EPAM Systems
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- ***************************************************************************/
-
-import styles from './PolymerToggler.module.css';
-
-export interface PolymerTogglerProps {
-  toggle: (isEnabled: boolean) => void;
-}
-
-const PolymerToggler = ({ toggle }: PolymerTogglerProps): JSX.Element => {
-  return (
-    <label className={styles.switch} data-testid="polymer-toggler">
-      {/* eslint-disable jsx-a11y/label-has-associated-control */}
-      <input type="checkbox" onChange={(e) => toggle(e.target.checked)} />
-      <span className={styles.slider} />
-      {/* eslint-enable jsx-a11y/label-has-associated-control */}
-    </label>
-  );
-};
-
-export { PolymerToggler };
