--- example-separate-editors/src/PolymerToggler/index.ts
@@ -1,1 +1,0 @@
-export * from './PolymerToggler';
