--- packages/ketcher-macromolecules/src/components/ModeControl/index.ts
@@ -1,1 +1,0 @@
-export * from './ModeControl';
