+++ packages/ketcher-react/rollup.config.js
@@ -7,7 +7,6 @@ import copy from 'rollup-plugin-copy';
 import del from 'rollup-plugin-delete';
 import json from '@rollup/plugin-json';
 import nodeResolve from '@rollup/plugin-node-resolve';
-import path from 'path';
 import peerDepsExternal from 'rollup-plugin-peer-deps-external';
 import pkg from './package.json';
 import postcss from 'rollup-plugin-postcss';
@@ -53,13 +52,13 @@ const config = {
   input: pkg.source,
   output: [
     {
-      file: pkg.main,
+      dir: 'dist/cjs',
       exports: 'named',
       format: 'cjs',
       banner: license,
     },
     {
-      file: pkg.module,
+      dir: 'dist',
       exports: 'named',
       format: 'es',
       banner: license,
@@ -72,7 +71,7 @@ const config = {
     }),
     postcss({
       plugins: [autoprefixer({ grid: 'autoplace' })],
-      extract: path.resolve('dist/index.css'),
+      extract: 'index.css',
       minimize: isProduction,
       sourceMap: true,
       include: includePattern,
