+++ packages/ketcher-react/src/Editor.module.less
@@ -161,6 +161,14 @@
   }
 }
 
+.switchingLoader {
+  height: 100%;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
 :global {
   .blinking {
     /* stylelint-disable-next-line selector-pseudo-class-no-unknown */
