+++ packages/ketcher-react/src/script/ui/views/toolbars/ModeControl/index.ts
